Refactor callouts/flashes/responses

+++ frontend/public_html/src/Controller/MusicController.php
@@ -53,12 +53,8 @@ class MusicController extends AppController
 
     public function addTitles()
     {
-        $errorFunc = function ($message, $userID, $type = null, $class = null) {
-            $this->_updatePlaylists($type, $class, $userID);
-            return $this->response->withStatus(500)->withStringBody(__('An error occurred during addTitles: ') . __($message));
-        };
         if (env('SERVER_PORT') != 82) {
-            return $this->response->withStatus(403)->withStringBody('Forbidden');
+            return $this->response->withStatus(403);
         }
         $token = $this->request->getData('request_id');
         $title_ids = $this->request->getData('song_ids');
@@ -66,48 +62,60 @@ class MusicController extends AppController
         $message = $this->request->getData('message');
 
         if (!isset($token, $code)) {
-            return $this->response->withStatus(400)->withStringBody('Bad request');
+            return $this->response->withStatus(400);
         }
 
-        $titlesToPlaylistTable = TableRegistry::getTableLocator()->get('TitlesToPlaylists');
         $addTitleTable = TableRegistry::getTableLocator()->get('AddTitlesJobs');
         try {
-            $addTitle = $addTitleTable->get($token);
+            // Retrieve job entry and delete from database
+            $addTitleTable->delete($addTitle = $addTitleTable->get($token));
         } catch (RecordNotFoundException $e) {
-            return $this->response->withStatus(200)->withStringBody('OK');
+            // If there is no job (likely because a user has deleted it), we can ignore the callback
+            return $this->response->withStatus(200);
         }
         $userID = $addTitle->get('user_id');
 
         switch ($code) {
             case 0:
-                break;
+                $titlesToPlaylistTable = TableRegistry::getTableLocator()->get('TitlesToPlaylists');
+                $playlistTable = TableRegistry::getTableLocator()->get('Playlists');
+
+                $playlistID = $addTitle->get('playlist_id');
+                $playlistName = $playlistTable->find('all', ['conditions' => ['id' => $playlistID]])->select('name')->first()->get('name');
+
+                $incompleteTitles = false;
+                $attemptedTitleCount = 0;
+                foreach ($title_ids as $title_id) {
+                    $titlesToPlaylist = $titlesToPlaylistTable->newEntity();
+                    $titlesToPlaylist->set('title_id', $title_id);
+                    $titlesToPlaylist->set('playlist_id', $playlistID);
+                    $attemptedTitleCount++;
+                    if (!$titlesToPlaylistTable->save($titlesToPlaylist)) {
+                        $this->log("Error saving titles_to_playlists");
+                        $incompleteTitles = true;
+                    }
+                }
+
+                $titleCount = $titlesToPlaylistTable->find('all', ['conditions' => ['playlist_id' => $playlistID]])->count();
+                if ($incompleteTitles) {
+                    $type = 'warning';
+                    $message = $titleCount . ' out of ' . $attemptedTitleCount . 'have successfully been added to "' . $playlistName . '"';
+                } else {
+                    $type = 'success';
+                    $message = $titleCount . ' titles have been successfully loaded into "' . $playlistName . '"';
+                }
+                $this->_updatePlaylists();
+                $this->_updateTitles($playlistID);
+                $this->_flash($type, $message, $userID);
+                return $this->response->withStatus(200);
             default:
-                return $errorFunc('Unknown code: ' . $code, $userID, 'alert', $message);
+                $this->_flash('alert', $message, $userID);
+                return $this->response->withStatus(500);
         }
-        foreach ($title_ids as $title_id) {
-            $titlesToPlaylist = $titlesToPlaylistTable->newEntity();
-            $titlesToPlaylist->set('title_id', $title_id);
-            $titlesToPlaylist->set('playlist_id', $addTitle->get('playlist_id'));
-            if (!$titlesToPlaylistTable->save($titlesToPlaylist)) {
-                return $errorFunc('Error saving title_to_playlist', $userID);
-            }
-        }
-
-        $addTitleTable->delete($addTitle);
-
-        $playlistTable = TableRegistry::getTableLocator()->get('Playlists');
-        $playlistName = $playlistTable->find('all', ['conditions' => ['id' => $addTitle->get('playlist_id')]])->select('name')->first()->get('name');
-        $titleCount = $titlesToPlaylistTable->find('all', ['conditions' => ['playlist_id' => $addTitle->get('playlist_id')]])->count();
-        $this->_updatePlaylists('success', $titleCount . ' titles have been successfully loaded into "' . $playlistName . '"', $userID);
-        return $this->response->withStatus(200)->withStringBody('OK');
     }
 
     public function addPlaylist()
     {
-        $errorFunc = function ($message) {
-            $this->_updatePlaylists('alert', __('An error occurred during playlist creation: ') . __($message));
-            return $this->response->withStatus(500)->withStringBody('Interal Server Error');
-        };
         $name = $this->request->getQuery('name');
         if (!isset($name) || mb_strlen($name) < 1) {
             return $this->response->withStatus(400)->withStringBody('Bad request');
@@ -117,37 +125,40 @@ class MusicController extends AppController
         $playlist = $playlistTable->newEntity();
         $playlist->set('name', $name);
         $playlist = $playlistTable->save($playlist);
-        if ($playlist) {
-            $url = $this->request->getQuery('url');
-            if (isset($url) && mb_strlen($url) > 0) {
-                try {
+        $status = 500;
+        try {
+            if ($playlist) {
+                $url = $this->request->getQuery('url');
+                if (isset($url) && mb_strlen($url) > 0) {
                     $response = $this->Api->createTitles($url);
-                } catch (Exception $e) {
-                    return $errorFunc('Unable to connect to backend');
-                }
-
-                if ($response->getStatusCode() === 202) {
-                    $addTitleTable = TableRegistry::getTableLocator()->get('AddTitlesJobs');
-                    $addTitle = $addTitleTable->newEntity();
-                    $addTitle->set('backend_token', $response->json['request_id']);
-                    $addTitle->set('playlist_id', $playlist->get('id'));
-                    $addTitle->set('user_id', $this->Auth->user('id'));
-                    if ($addTitleTable->save($addTitle)) {
-                        $this->_updatePlaylists();
-                        return $this->response->withStatus(200)->withStringBody('Your playlist is now in processing. You will be notified once it is fully loaded');
+                    if ($response->getStatusCode() === 202) {
+                        $addTitleTable = TableRegistry::getTableLocator()->get('AddTitlesJobs');
+                        $addTitle = $addTitleTable->newEntity();
+                        $addTitle->set('backend_token', $response->getJson()['request_id']);
+                        $addTitle->set('playlist_id', $playlist->get('id'));
+                        $addTitle->set('user_id', $this->Auth->user('id'));
+                        if ($addTitleTable->save($addTitle)) {
+                            $status = 200;
+                            $message = 'Your playlist is now in processing. You will be notified once it is fully loaded';
+                        } else {
+                            $message = 'Database Error';
+                        }
                     } else {
-                        return $errorFunc('Unable to create addTitleJob');
+                        $message = 'Could not resolve URL';
                     }
                 } else {
-                    return $errorFunc('Could not resolve URL');
+                    $status = 200;
+                    $message = 'OK';
                 }
             } else {
-                $this->_updatePlaylists();
-                return $this->response->withStatus(200)->withStringBody('OK');
+                $message = 'Could not save the playlist';
             }
-        } else {
-            return $errorFunc('Could not save the playlist');
+        } catch (Exception $e) {
+            $message = 'Unable to connect to backend';
         }
+
+        $this->_updatePlaylists();
+        return $this->response->withStatus($status)->withStringBody($message);
     }
 
     private function _playlistsJson()
@@ -181,20 +192,23 @@ class MusicController extends AppController
     {
         $id = $this->request->getQuery('id');
         if (!isset($id) || mb_strlen($id) !== 36) {
-            return $this->response->withStatus(400)->withStringBody('Bad request');
+            return $this->response->withStatus(400);
         }
         $playlistTable = TableRegistry::getTableLocator()->get('Playlists');
         $playlist = $playlistTable->get($id);
 
-        try {
-            $this->Api->deleteTitles($playlist->get('id'));
-        } catch (Exception $e) {
-            return $this->response->withStatus(500)->withStringBody('Unable to connect to backend');
+        $titlesToPlaylistTable = TableRegistry::getTableLocator()->get('TitlesToPlaylists');
+        if ($titlesToPlaylistTable->find()->where(['playlist_id' => $id])->count()) {
+            try {
+                $this->Api->deleteTitles($playlist->get('id'));
+            } catch (Exception $e) {
+                return $this->response->withStatus(500)->withStringBody('Unable to connect to backend');
+            }
         }
 
         $playlistTable->delete($playlist);
         $this->_updatePlaylists();
-        return $this->response->withStatus(200)->withStringBody('OK');
+        return $this->response->withStatus(200);
     }
 
     public function deleteTitle($playlist_id, $title_id)
@@ -210,7 +224,7 @@ class MusicController extends AppController
         }
         $this->_updatePlaylists();
         $this->_updateTitles($playlist_id);
-        return $this->response->withStatus(200)->withStringBody('OK');
+        return $this->response->withStatus(200);
     }
 
     private function _updateTitles($playlist_id)
@@ -229,11 +243,16 @@ class MusicController extends AppController
         return $this->response->withType('json')->withStringBody($this->_titlesJson($playlist_id));
     }
 
-    private function _updatePlaylists($type = null, $message = null, $userID = null)
+    private function _updatePlaylists()
+    {
+        Websocket::publishEvent('playlistsUpdated', ['json' => $this->_playlistsJson()]);
+    }
+
+    private function _flash($type = null, $message = null, $userID = null)
     {
         if ($userID == null) {
             $userID = $this->Auth->user('id');
         }
-        Websocket::publishEvent('playlistsUpdated', ['json' => $this->_playlistsJson(), 'type' => $type, 'message' => $message, 'userID' => $userID]);
+        Websocket::publishEvent('flash', ['type' => $type, 'message' => $message, 'userID' => $userID]);
     }
 }
